Can Now interface the arduino with Berry IMU
The interface for the GPS and Berry IMU is finished! The arduino can communicate between the Berry IMU I2C bus and the GPS module at the same time! All that needs to be done now is to add a serial interface for the laptop to communicate with the arduino and it's finished!