Potentiometer Test Code FInished
The code for measuring the angle of the front wheels is finished. Features include:
Autoscaling: the program will figure out where the maximum and minimum value of the potentiometer is by itself.

Value filtering: Often times the potentiomter can return a sudden spike in its value which htrows off the auto scaling. This no longer happens.