Adjusted Program so it can do more than one thing at a time
It's often time consuming to have the arduino micromanage the steering wheel until it gets to the target position. During this time the arduino can't send or recieve any commands from or to the laptop. With the new setup, you can have the sterring wheel get in place while also sending and recieving commands.