Movement Control Updated NEEDS TESTING BEFORE PROCEEDING
I changed up the arduino movement control sketch so that it actually makes sense now. I also implemented the call and response handshake between the arduino and and laptop so they know they're communicating with each other. I also enabled the arduino to control the steering wheel and do other things at the same time. This may add soem issues to the steering but we'll see. Finally, I also added a function in the python class to handle the arduino communication and I'm to test it out when I have access to an arduino. Finally I removed unecessary arduino sketches like pot_test, and replaced movementcontrol with movementControl2.